Fix all compiler warnings in thread_system build
- Fix sign conversion warnings in convert_string.cpp by adding explicit casts
- Fix constructor initialization order warnings in multiple files:
  - callback_job.cpp: Reorder member initialization to match declaration order
  - thread_base.cpp: Reorder wake_interval_ initialization
  - message_job.cpp: Fix member initialization order
  - file_writer.cpp: Correct initialization sequence
  - log_job.cpp: Match declaration order for message type members
  - log_collector.cpp: Fix queue and log type initialization order
  - thread_worker.cpp: Reorder use_time_tag_ and job_queue_ initialization
  - thread_pool.cpp: Fix start_pool_ and job_queue_ initialization order
- Fix redundant move warnings in file_writer.cpp by removing unnecessary std::move
- Fix sign conversion warnings in file_writer.cpp with explicit streamsize casts
- Suppress duplicate library warnings on macOS with -Wl,-no_warn_duplicate_libraries flag

+++ sources/utilities/convert_string.cpp
@@ -452,7 +452,7 @@ namespace utility_module
 		int mod_table[] = { 0, 2, 1 };
 		for (int j = 0; j < mod_table[data.size() % 3]; j++)
 		{
-			encoded_string[encoded_string.size() - 1 - j] = '=';
+			encoded_string[encoded_string.size() - 1 - static_cast<size_t>(j)] = '=';
 		}
 
 		return encoded_string;
@@ -490,7 +490,7 @@ namespace utility_module
 		std::vector<int> decoding_table(256, -1);
 		for (int i = 0; i < 64; i++)
 		{
-			decoding_table[static_cast<unsigned char>(base64_chars[i])] = i;
+			decoding_table[static_cast<unsigned char>(base64_chars[static_cast<size_t>(i)])] = i;
 		}
 
 		uint32_t buffer = 0;
@@ -513,7 +513,7 @@ namespace utility_module
 				return { std::vector<uint8_t>(), "Invalid character in base64 string" };
 			}
 
-			buffer = (buffer << 6) | decoding_table[static_cast<unsigned char>(c)];
+			buffer = (buffer << 6) | static_cast<uint32_t>(decoding_table[static_cast<unsigned char>(c)]);
 			bits_collected += 6;
 
 			if (bits_collected >= 8)
